chore: separated bin and lib crates
