+++ crates/rsonpath-benchmarks/Cargo.toml
@@ -24,13 +24,8 @@ jni = { version = "0.19.0", features = ["invocation", "default"] }
 libc = "0.2.126"
 lazy_static = "1.4.0"
 ouroboros = "0.15.5"
-rsonpath = { version = "0.1.2", default-features = false }
+rsonpath = { package = "rsonpath-lib", path = "../rsonpath-lib", default-features = false }
 thiserror = "1.0.37"
-
-[dev-dependencies]
-itertools = "0.10.3"
-regex = "1.6.0"
-memchr = "2.5.0"
 
 [features]
 default = ["simd"]
