+++ crates/rsonpath-lib/src/classify/avx2.rs
