+++ crates/rsonpath-lib/src/classify/nosimd.rs
