+++ crates/rsonpath-lib/src/depth.rs
