+++ crates/rsonpath-lib/src/depth/avx2.rs
