+++ crates/rsonpath-lib/src/depth/nosimd.rs
