+++ crates/rsonpath-lib/src/engine/result.rs
