+++ crates/rsonpath-lib/src/query/automaton/minimizer.rs
