+++ crates/rsonpath-lib/src/query/automaton/superstate.rs
