+++ crates/rsonpath-lib/src/query/errors.rs
