+++ crates/rsonpath-lib/src/stack_based.rs
