+++ crates/rsonpath-lib/src/stackless.rs
