+++ crates/rsonpath-lib/tests/data/basic/child_hell.json
