+++ crates/rsonpath-lib/tests/data/basic/compressed/child.json
