+++ crates/rsonpath-lib/tests/data/basic/compressed/child_hell.json
