+++ crates/rsonpath-lib/tests/data/basic/compressed/escapes.json
