+++ crates/rsonpath-lib/tests/data/basic/compressed/quote_escape.json
