+++ crates/rsonpath-lib/tests/data/basic/compressed/root.json
