+++ crates/rsonpath-lib/tests/data/basic/compressed/skipping.json
