+++ crates/rsonpath-lib/tests/data/basic/compressed/small.json
