+++ crates/rsonpath-lib/tests/data/basic/compressed/small_no_list.json
