+++ crates/rsonpath-lib/tests/data/basic/compressed/spaced_colon.json
