+++ crates/rsonpath-lib/tests/data/basic/small_no_list.json
