+++ crates/rsonpath-lib/tests/data/basic/spaced_colon.json
