+++ crates/rsonpath-lib/tests/data/twitter/compressed/twitter.json
