+++ crates/rsonpath-lib/tests/data/twitter/compressed/twitter_urls.json
